15 juni 2022

+++ detail_preprocess.py
@@ -34,7 +34,6 @@ def btn_fliplr_image():
 
     # print(os.path.exists(file_path_variable1))
     if (os.path.exists(file_path_variable1) and os.path.exists(file_path_variable2)) is True:
-        print("yaa")
 
         images = []
         images_path = glob.glob(file_path_variable1 + "/*.*")
@@ -45,7 +44,6 @@ def btn_fliplr_image():
         augmentation = iaa.Sequential([
             iaa.Fliplr(1.0)
         ])
-
         # Show and Save Images
         augmented_images = augmentation(images=images)
 
@@ -59,15 +57,13 @@ def btn_fliplr_image():
             i +=1
             sum_img = x + i
             
-            if cv2.waitKey(1) & 0xFF == ord('q'):
-                break
-            # cv2.waitKey(0)
+            cv2.waitKey(5)
 
         messagebox.showinfo("Show Info", "Gambar Berhasil di-flip LR!")
         print("Total Image Setelah di Flip LR: ", sum_img)
         cv2.destroyAllWindows()
     else:
-        print("jala")
+        messagebox.showinfo("Show Info", "Path harus ada!")
 
 # fungsi flip ud image
 def btn_flipud_image():
@@ -75,36 +71,40 @@ def btn_flipud_image():
     file_path_variable1 = filedialog.askdirectory(parent=root, initialdir=currdir, title='Please select a directory')
     file_path_variable2 = filedialog.askdirectory(parent=root, initialdir=currdir, title='Please select a directory to Save')
 
-    images = []
-    images_path = glob.glob(file_path_variable1 + "/*.jpg")
-    for img_path in images_path:
-        img = cv2.imread(img_path)
-        images.append(img)
+    if (os.path.exists(file_path_variable1) and os.path.exists(file_path_variable2)) is True:
 
-    augmentation = iaa.Sequential([
-        iaa.Flipud(1.0)
-    ])
+        images = []
+        images_path = glob.glob(file_path_variable1 + "/*.jpg")
+        for img_path in images_path:
+            img = cv2.imread(img_path)
+            images.append(img)
 
-    # Show and Save Images
-    augmented_images = augmentation(images=images)
+        augmentation = iaa.Sequential([
+            iaa.Flipud(1.0)
+        ])
 
-    i=0
-    y=0
-    for img in augmented_images:
-        cv2.imshow("Image", img)
+        # Show and Save Images
+        augmented_images = augmentation(images=images)
+
+        i=0
+        y=0
+        for img in augmented_images:
+            cv2.imshow("Image", img)
+            
+            cv2.imwrite(file_path_variable2 + "/citra%03i.jpg" %i, img)
+            
+            i +=1
+            sum_img = y + i
+
+            if cv2.waitKey(1) & 0xFF == ord('q'):
+                break
+            # cv2.waitKey(0)
         
-        cv2.imwrite(file_path_variable2 + "/citra%03i.jpg" %i, img)
-        
-        i +=1
-        sum_img = y + i
-
-        if cv2.waitKey(1) & 0xFF == ord('q'):
-            break
-        # cv2.waitKey(0)
-    
-    messagebox.showinfo("Show Info", "Gambar Berhasil di-flip UD!")
-    print("Total Image Setelah di Flip UD: ", sum_img)
-    cv2.destroyAllWindows()
+        messagebox.showinfo("Show Info", "Gambar Berhasil di-flip UD!")
+        print("Total Image Setelah di Flip UD: ", sum_img)
+        cv2.destroyAllWindows()
+    else:
+        messagebox.showinfo("Show Info", "Path harus ada")
 
 # fungsi rotasi image
 def btn_rotate_image():
@@ -112,37 +112,40 @@ def btn_rotate_image():
     file_path_variable1 = filedialog.askdirectory(parent=root, initialdir=currdir, title='Please select a directory')
     file_path_variable2 = filedialog.askdirectory(parent=root, initialdir=currdir, title='Please select a directory to Save')
 
-    images = []
-    images_path = glob.glob(file_path_variable1 + "/*.*")
-    for img_path in images_path:
-        img = cv2.imread(img_path)
-        images.append(img)
+    if (os.path.exists(file_path_variable1) and os.path.exists(file_path_variable2)) is True:
+        images = []
+        images_path = glob.glob(file_path_variable1 + "/*.*")
+        for img_path in images_path:
+            img = cv2.imread(img_path)
+            images.append(img)
 
-    augmentation = iaa.Sequential([
-        iaa.Affine(translate_px={"x": (-20, 20), "y": (-20, 20)},
-            rotate=(-25, 25))
-    ])
+        augmentation = iaa.Sequential([
+            iaa.Affine(translate_px={"x": (-20, 20), "y": (-20, 20)},
+                rotate=(-25, 25))
+        ])
 
-    # 3. Show and Save Images
-    augmented_images = augmentation(images=images)
+        # 3. Show and Save Images
+        augmented_images = augmentation(images=images)
 
-    i=0
-    z=0
-    for img in augmented_images:
-        cv2.imshow("Image", img)
-        
-        cv2.imwrite(file_path_variable2 + "/gbr%03i.jpg" %i, img)
+        i=0
+        z=0
+        for img in augmented_images:
+            cv2.imshow("Image", img)
+            
+            cv2.imwrite(file_path_variable2 + "/gbr%03i.jpg" %i, img)
 
-        i +=1
-        sum_img = z + i
+            i +=1
+            sum_img = z + i
 
-        if cv2.waitKey(1) & 0xFF == ord('q'):
-            break
-        # cv2.waitKey(0)
+            if cv2.waitKey(1) & 0xFF == ord('q'):
+                break
+            # cv2.waitKey(0)
 
-    messagebox.showinfo("Show Info", "Gambar Berhasil dirotasi!")
-    print("Total Image Setelah Dirotasi: ", sum_img)
-    cv2.destroyAllWindows()
+        messagebox.showinfo("Show Info", "Gambar Berhasil dirotasi!")
+        print("Total Image Setelah Dirotasi: ", sum_img)
+        cv2.destroyAllWindows()
+    else:
+        messagebox.showinfo("Show Info", "Path harus ada!")
 
 # fungsi zoom images
 def btn_zoomin_image():
@@ -150,37 +153,40 @@ def btn_zoomin_image():
     file_path_variable1 = filedialog.askdirectory(parent=root, initialdir=currdir, title='Please select a directory')
     file_path_variable2 = filedialog.askdirectory(parent=root, initialdir=currdir, title='Please select a directory to Save')
 
-    images = []
-    images_path = glob.glob(file_path_variable1 + "/*.*")
-    for img_path in images_path:
-        img = cv2.imread(img_path)
-        images.append(img)
+    if (os.path.exists(file_path_variable1) and os.path.exists(file_path_variable2)) is True:
+        images = []
+        images_path = glob.glob(file_path_variable1 + "/*.*")
+        for img_path in images_path:
+            img = cv2.imread(img_path)
+            images.append(img)
 
-    augmentation = iaa.Sequential([
-        iaa.Affine(
-        scale={"x": (0.4, 1.0), "y": (0.8, 1.2)})
-    ])
+        augmentation = iaa.Sequential([
+            iaa.Affine(
+            scale={"x": (0.4, 1.0), "y": (0.8, 1.2)})
+        ])
 
-    # 3. Show and Save Images
-    augmented_images = augmentation(images=images)
+        # 3. Show and Save Images
+        augmented_images = augmentation(images=images)
 
-    i=0
-    a=0
-    for img in augmented_images:
-        cv2.imshow("Image", img)
+        i=0
+        a=0
+        for img in augmented_images:
+            cv2.imshow("Image", img)
+            
+            cv2.imwrite(file_path_variable2 + "/ctr%03i.jpg" %i, img)
+
+            i +=1
+            sum_img = a + i
+
+            if cv2.waitKey(1) & 0xFF == ord('q'):
+                break
+            # cv2.waitKey(0)
         
-        cv2.imwrite(file_path_variable2 + "/ctr%03i.jpg" %i, img)
-
-        i +=1
-        sum_img = a + i
-
-        if cv2.waitKey(1) & 0xFF == ord('q'):
-            break
-        # cv2.waitKey(0)
-    
-    messagebox.showinfo("Show Info", "Gambar Berhasil di-zoom!")
-    print("Total Image Setelah Di-zoom: ", sum_img)
-    cv2.destroyAllWindows()
+        messagebox.showinfo("Show Info", "Gambar Berhasil di-zoom!")
+        print("Total Image Setelah Di-zoom: ", sum_img)
+        cv2.destroyAllWindows()
+    else:
+        messagebox.showinfo("Show Info", "Path harus ada!")
 
 # fungsi memproses edge detection
 def btn_edge_detect():
@@ -191,27 +197,31 @@ def btn_edge_detect():
     file_path_variable1 = filedialog.askdirectory(parent=root, initialdir=currdir, title='Please Select a Directory Raw Dataset 2')
     file_path_variable2 = filedialog.askdirectory(parent=root, initialdir=currdir, title='Please Select a Directory to Save in Fix Dataset')
     
-    for img in glob.glob(file_path_variable1 + "/*.*"):
-        image = cv2.imread(img)
-        ImgResized = cv2.resize(image, (224, 224))
+    if (os.path.exists(file_path_variable1) and os.path.exists(file_path_variable2)) is True:
+        for img in glob.glob(file_path_variable1 + "/*.*"):
+            image = cv2.imread(img)
+            ImgResized = cv2.resize(image, (224, 224))
 
-        # convert RGB to Grayscale image
-        ImgGray = cv2.cvtColor(ImgResized, cv2.COLOR_BGR2GRAY)
-        # canny edge detection
-        ImgEdge = cv2.Canny(ImgGray, 100, 200)
+            # convert RGB to Grayscale image
+            ImgGray = cv2.cvtColor(ImgResized, cv2.COLOR_BGR2GRAY)
+            # canny edge detection
+            ImgEdge = cv2.Canny(ImgGray, 100, 200) #jabarkan -> scratch
 
-        # save image in custom folder
-        cv2.imwrite(file_path_variable2 + "/image%03i.jpg" %i, ImgEdge)
+            # save image in custom folder
+            cv2.imwrite(file_path_variable2 + "/image%03i.jpg" %i, ImgEdge)
 
-        i +=1
-        sum_img = j + i
-        
-        cv2.imshow('image', ImgEdge)
-        cv2.waitKey(10)
+            i +=1
+            sum_img = j + i
+            
+            cv2.imshow('image', ImgEdge)
+            cv2.waitKey(10)
 
-    messagebox.showinfo("Show Info", "Gambar Berhasil di-deteksi tepi. Ready to use!")
-    print("Total keseluruhan image (ready):", sum_img)
-    cv2.destroyAllWindows()
+        messagebox.showinfo("Show Info", "Gambar Berhasil di-deteksi tepi. Ready to use!")
+        print("Total keseluruhan image (ready):", sum_img)
+        print(type(ImgEdge))
+        cv2.destroyAllWindows()
+    else:
+        messagebox.showinfo("Show Info", "Path harus ada!")
 
 # button control
 frame = Frame(root, bg='#F2B33D')
